+++ .github/workflows/quiz-batch-modul.yml
@@ -4,7 +4,7 @@ on:
   pull_request:
     branches: 
  #     - main
-      - release/v1.0.2
+      - release/v1.0.3
     paths:
       - 'spring-module/quizstudio/quiz-batch/**'
